Blue squad test directory re-structure (#8999)
* change directory structure for blue squad test cases


* distribute blue squad tests to libtest and workload folders


* fix flake8 errors


* add missing threading_lock


* fix bad variable name


---------
