+++ tests/functional/monitoring/libtest/test_workload_fixture.py
@@ -43,8 +43,8 @@ from datetime import datetime
 
 import pytest
 
-from ocs_ci.framework.pytest_customization.marks import blue_squad, skipif_mcg_only
-from ocs_ci.framework.testlib import tier1, skipif_managed_service
+from ocs_ci.framework.pytest_customization.marks import blue_squad
+from ocs_ci.framework.testlib import skipif_managed_service
 from ocs_ci.utility.prometheus import PrometheusAPI
 
 
@@ -179,43 +179,3 @@ def test_workload_rbd_cephfs_minimal(
     """
     logger.info(workload_storageutilization_05p_rbd)
     logger.info(workload_storageutilization_05p_cephfs)
-
-
-@blue_squad
-@skipif_mcg_only
-@tier1
-@pytest.mark.polarion_id("OCS-2125")
-@skipif_managed_service
-def test_workload_rbd_cephfs_10g(
-    workload_storageutilization_10g_rbd, workload_storageutilization_10g_cephfs
-):
-    """
-    Test of a workload utilization with constant 10 GiB target.
-
-    In this test we are only checking whether the storage utilization workload
-    failed or not. The main point of having this included in tier1 suite is to
-    see whether we are able to actually run the fio write workload without any
-    direct failure (fio job could fail to be scheduled, fail during writing or
-    timeout when write progress is too slow ...).
-    """
-    logger.info("checking fio report results as provided by workload fixtures")
-    msg = "workload results should be recorded and provided to the test"
-    assert workload_storageutilization_10g_rbd["result"] is not None, msg
-    assert workload_storageutilization_10g_cephfs["result"] is not None, msg
-
-    fio_reports = (
-        ("rbd", workload_storageutilization_10g_rbd["result"]["fio"]),
-        ("cephfs", workload_storageutilization_10g_cephfs["result"]["fio"]),
-    )
-    for vol_type, fio in fio_reports:
-        logger.info("starting to check fio run on %s volume", vol_type)
-        msg = "single fio job should be executed in each workload run"
-        assert len(fio["jobs"]) == 1, msg
-        logger.info(
-            "fio (version %s) executed %s job on %s volume",
-            fio["fio version"],
-            fio["jobs"][0]["jobname"],
-            vol_type,
-        )
-        msg = f"no errors should be reported by fio writing on {vol_type} volume"
-        assert fio["jobs"][0]["error"] == 0, msg
