+++ tests/functional/monitoring/prometheus/metrics/test_mcg_hpa.py
