+++ tests/functional/monitoring/prometheus/metrics/test_monitoring_defaults.py
