+++ tests/functional/monitoring/prometheus/metrics/test_rgw.py
